Move user-manuals to install-manuals and add real user-manuals (#154)
Part of Gamify-IT/issues#309

### Contents of this PR
- Move the contents of `user-manuals/` to `install-manuals/`
- Create an actual user manual (meant for lecturers or students) for
chickenshock

### Reasoning
What we currently have in the `user-manuals/` folder is only of concern
for people hosting our software. Therefore I renamed
the directory `user-manuals/` to `install-manuals/`.

In our code review, we got the feedback, that we are missing high-level,
non-technical documentation for people unrelated to the project. I want
to repurpose the `user-manuals/` directory to collect documentation for
our **actual users**, meaning **lecturers and students**. The
documentation should describe, how the overworld, minigames, and lecturer
interface work (how to play the games and how to use the UI). It should
contain screenshots and useful diagrams. We can then reference this
documentation from the minigame repository README.md files and other
resources.

Co-authored-by: delvh <leon@kske.dev>
Co-authored-by: LeviUniAcc <102458061+LeviUniAcc@users.noreply.github.com>
Co-authored-by: delvh <dev.lh@web.de>